+++ Scripts/Global/SkillEffect.lua
@@ -1,12 +1,94 @@
 function events.GetAttackDelay(t)
 
-	-- get recovery time bonus
-	local recoveryTimeBonus = 100 - t.Result
-	
 	-- correct weapon skill recovery time bonus
 	
+	if t.Ranged then
+	else
+	
+		-- find recovery time skill
+	
+		local recoveryTimeSkill = nil
+		
+		local itemMainHandNumber = t.Player.ItemMainHand
+		if itemMainHandNumber ~= 0 then
+			local itemMainHand = t.Player.Items[itemMainHandNumber]
+			local itemMainHandSkillInc = Game.ItemsTxt[itemMainHand.Number].Skill
+			
+			if itemMainHandSkillInc ~= 0 then
+				local itemMainHandSkill = itemMainHandSkillInc - 1
+				
+				local itemExtraHandNumber = t.Player.ItemExtraHand
+				if itemExtraHandNumber == 0 then
+					recoveryTimeSkill = itemMainHandSkill
+				else
+					local itemExtraHand = t.Player.Items[itemExtraHandNumber]
+					local itemExtraHandSkillInc = Game.ItemsTxt[itemExtraHand.Number].Skill
+					
+					if itemExtraHandSkillInc == 0 then
+						recoveryTimeSkill = itemMainHandSkill
+					else
+						
+						local itemExtraHandSkill = itemExtraHandSkillInc - 1
+						
+						if itemMainHandSkill == const.Skills.Axe then
+							recoveryTimeSkill = const.Skills.Axe
+						elseif itemMainHandSkill == const.Skills.Sword or itemExtraHandSkill == const.Skills.Sword then
+							recoveryTimeSkill = const.Skills.Sword
+						else
+							recoveryTimeSkill = nil
+						end
+						
+					end
+					
+				end
+				
+			end
+			
+		end
+		
+		if recoveryTimeSkill ~= nil then
+			
+			local itemExtraHandNumber = t.Player.ItemExtraHand
+			if itemExtraHandNumber ~= 0 then
+				local itemExtraHand = t.Player.Items[itemExtraHandNumber]
+				local itemExtraHandSkillInc = Game.ItemsTxt[itemExtraHand.Number].Skill
+				
+				if itemExtraHandSkillInc ~= 0 then
+					local itemExtraHandSkill = itemExtraHandSkillInc - 1
+					
+					if	 	itemExtraHandSkill == const.Skills.Sword	then
+						if baseRecoveryTime < 90 then
+							recoveryTimeSkill = itemExtraHandSkill
+							baseRecoveryTime = 90
+						end
+					end
+					
+				end
+				
+			end
+			
+		end
+		
+		-- calculate recovery time adjustment
+		
+		if recoveryTimeSkill ~= nil then
+			
+			if recoveryTimeSkill == const.Skills.Sword or recoveryTimeSkill == const.Skills.Axe then
+				if t.Player.Skills[recoveryTimeSkill] ~= 0 then
+					local skill, mastery = SplitSkill(t.Player.Skills[recoveryTimeSkill])
+					if mastery >= const.Expert then
+						t.Result = t.Result - (skill * 9)
+					end
+				end
+			end
+			
+		end
+	
+	end
 	
 	-- turn recovery time into a multiplier rather than divisor
+	
+	local recoveryTimeBonus = 100 - t.Result
 	local correctedRecoveryTime = math.floor(100 * 100 / (100 + recoveryTimeBonus))
 	t.Result = correctedRecoveryTime