+++ Changelog.md
@@ -1,4 +1,4 @@
 # version 1
 
 * Recovery time increases multiplier instead of decreasing divider.
-
+* Expert Axe and Sword skill recovery time bonus is 10.